+++ lib/events.rb
@@ -19,11 +19,11 @@ SlackRubyBotServer::Events.configure do |config|
     next { ok: false } unless data
 
     Api::Middleware.logger.info "#{data.team.name}: bot joined ##{data.channel}."
-    data.team.join_channel!(data.channel, data.inviter)
+    channel = data.team.join_channel!(data.channel, data.inviter)
 
     text = [
       "Hi there! I'm your team's S'Up bot.",
-      "I will sync users in this channel in the next hour, and plan to setup a S'Up via Slack DM for all users in this channel next Monday.",
+      "I will sync users in this channel in the next hour, and plan to setup a S'Up via Slack DM for all users in this channel tomorrow, #{channel.sup_day}.",
       "You may want to `#{data.team.bot_name} set size`, `#{data.team.bot_name} set day`, or `#{data.team.bot_name} set timezone` before then.",
       "Type `#{data.team.bot_name} help` for instructions and `#{data.team.bot_name} stats` for S'Up info."
     ].join(' ')